requirements.txt
anyio==3.6.2
asgiref==3.4.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.0.1
distlib==0.3.6
dnspython==1.16.0
fastapi==0.70.0
filelock==3.9.0
Flask==2.2.2
h11==0.12.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
platformdirs==2.6.2
pydantic==1.8.2
PyJWT==1.7.1
pymongo==4.0.1
python-decouple==3.3
python-dotenv==0.21.1
qrcode==7.3.1
requests==2.28.2
sniffio==1.3.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.14
uvicorn==0.15.0
virtualenv==20.17.1
Werkzeug==2.2.2
